Migrate buildsystem from Premake to Meson
Premake was more trouble than it was worth. We are using meson on our other
projects, and it provides a simple interface for defining builds. It's also portable
across systems and is easily installed by the user.

Because the build submodule changed, users may need to re-clone the repository.

At the very least, remove the `build/` folder, perform a `git submodule sync`,
and re-download the submodule.

Instructions for installing meson are included in the README.

Fixes #22